pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "studiosync-core"
version = "0.1.0"
description = "A comprehensive ERP suite for music studios, including billing, scheduling, and student management."
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "Django>=5.0.0",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.3.0",
    "django-filter>=24.0",
    "python-dotenv>=1.0.0",
    "django-q2>=1.7.0",
    "psycopg2-binary>=2.9.0",
    "daphne>=4.1.0",
    "channels>=4.0.0",
    "djangorestframework-simplejwt>=5.3.0",
    "django-allauth>=0.60.0",
    "django-storages>=1.14.0",
    "boto3>=1.35.0",
    "Pillow>=10.4.0",
    "stripe>=11.5.0",
    "drf-spectacular>=0.28.0",
    "whitenoise>=6.6.0",
    "dj-database-url>=2.1.0",
    "caldav>=1.3.0",
    "icalendar>=5.0.0",
    "python-magic>=0.4.27",
    "bleach>=6.1.0",
    "markdown>=3.5.0",
    "stream-chat"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-django>=4.9.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.25.0",
    "factory-boy>=3.3.0",
    "faker>=40.0.0",
    "black>=24.0.0",
    "ruff>=0.14.0",
    "isort>=5.13.0",
    "pre-commit>=4.0.0",
    "moto>=5.1.0"
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["studiosync_core*"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = ["**/migrations/**"]
